Fix for backend changes, evade GFW
Twitch changed the TLS certificate's alt names, breaking the evasion I was using for the GFW. Fix by accepting any names on the certificate.

Dangerous in the general case, but here the worst thing CNGov can do is make you watch a different video, which was already possible if they hijacked your function.